Move Checker Function
